북마크리스트 개발완료

+++ loggalApp/app/src/main/java/com/altsoft/model/MOBILE_AD_DETAIL_COND.java
@@ -8,7 +8,7 @@ import java.math.BigInteger;
 /// </summary>
 public class MOBILE_AD_DETAIL_COND implements Serializable {
     /// 배너코드
-    public BigInteger AD_CODE;
+    public Long AD_CODE;
     /// 사용자아이디(이메일)
     public String USER_ID;
 }
