+++ loggalApp/app/src/main/java/com/altsoft/model/T_AD.java
@@ -5,7 +5,7 @@ import java.math.BigInteger;
 
 public class T_AD implements Serializable  {
     /// 광고코드
-    public BigInteger AD_CODE;
+    public Long AD_CODE;
     /// 광고배너표시유형(T_COMMON : A005) (1:제목+이미지, 2:제목,3:이미지)
     public Integer BANNER_TYPE;
     /// 콘텐츠 유형(T_COMMON : A008, 1:HTML, 2:서브배너)
